Реализован модуль Payment

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/order/OrderDto.java
@@ -3,7 +3,6 @@ package ru.yandex.practicum.dto.order;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 
@@ -21,7 +20,7 @@ public class OrderDto {
     private Double deliveryWeight;
     private Double deliveryVolume;
     private Boolean fragile;
-    private BigDecimal totalPrice;
-    private BigDecimal deliveryPrice;
-    private BigDecimal productPrice;
+    private Double totalPrice;
+    private Double deliveryPrice;
+    private Double productPrice;
 }
