+++ commerce/order/src/main/java/ru/yandex/practicum/model/Order.java
@@ -17,7 +17,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import ru.yandex.practicum.dto.order.OrderState;
 
-import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 
@@ -53,9 +52,9 @@ public class Order {
     private Double deliveryVolume;
     private Boolean fragile;
     @Column(name = "total_price")
-    private BigDecimal totalPrice;
+    private Double totalPrice;
     @Column(name = "delivery_price")
-    private BigDecimal deliveryPrice;
+    private Double deliveryPrice;
     @Column(name = "product_price")
-    private BigDecimal productPrice;
+    private Double productPrice;
 }
